backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
supabase==2.8.1
pdfplumber==0.10.3
pandas==2.1.4
openpyxl==3.1.2
python-multipart==0.0.6
httpx==0.24.1


